+++ src/components/RenderRoles/RenderRoles.js
@@ -7,13 +7,15 @@ import {
   Accordion,
   Badge,
   Headline,
-  Loading
+  Loading,
+  TextLink,
 } from '@folio/stripes/components';
 
 import AffiliationsSelect from '../AffiliationsSelect/AffiliationsSelect';
 import IfConsortium from '../IfConsortium';
 import IfConsortiumPermission from '../IfConsortiumPermission';
 import { affiliationsShape } from '../../shapes';
+import { getRoleDetailPath } from '../util/util';
 
 class RenderRoles extends React.Component {
   static propTypes = {
@@ -49,7 +51,9 @@ class RenderRoles extends React.Component {
     const {
       listedRoles,
     } = this.props;
-    const listFormatter = item => <li key={item.id}>{item.name}</li>;
+    const listFormatter = item => <li key={item.id}>
+      <TextLink to={getRoleDetailPath(item.id)} target="_blank">{item.name}</TextLink>
+    </li>;
     const noPermissionsFound = <FormattedMessage id="ui-users.roles.empty" />;
 
     return (